move the `core/utils` inside the `core/base`
