+++ hydrolib/core/dflowfm/ext/models.py
@@ -22,7 +22,7 @@ from hydrolib.core.dflowfm.ini.util import (
 )
 from hydrolib.core.dflowfm.polyfile.models import PolyFile
 from hydrolib.core.dflowfm.tim.models import TimModel
-from hydrolib.core.utils import str_is_empty_or_none
+from hydrolib.core.base.utils import str_is_empty_or_none
 
 SOURCE_SINKS_QUANTITIES_VALID_PREFIXES = (
     "initialtracer",
