+++ hydrolib/core/dflowfm/tim/models.py
@@ -8,7 +8,7 @@ from pydantic.v1.class_validators import validator
 from hydrolib.core.base.models import BaseModel, ModelSaveSettings, ParsableFileModel
 from hydrolib.core.dflowfm.tim.parser import TimParser
 from hydrolib.core.dflowfm.tim.serializer import TimSerializer, TimSerializerConfig
-from hydrolib.core.utils import FortranUtils
+from hydrolib.core.base.utils import FortranUtils
 
 
 class TimRecord(BaseModel):
