+++ hydrolib/core/dimr/models.py
@@ -19,7 +19,7 @@ from hydrolib.core.dflowfm.mdu.models import FMModel
 from hydrolib.core.dimr.parser import DIMRParser
 from hydrolib.core.dimr.serializer import DIMRSerializer
 from hydrolib.core.rr.models import RainfallRunoffModel
-from hydrolib.core.utils import to_list
+from hydrolib.core.base.utils import to_list
 
 
 class KeyValuePair(BaseModel):
