+++ hydrolib/core/rr/serializer.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Dict, Iterable, Optional, Union
 
 from hydrolib.core.base.models import ModelSaveSettings, SerializerConfig
-from hydrolib.core.utils import FilePathStyleConverter, get_str_len
+from hydrolib.core.base.utils import FilePathStyleConverter, get_str_len
 
 
 def _calculate_max_value_length(data: Iterable) -> int:
